validation bulk, log membership

+++ app/Http/Controllers/AccountController.php
@@ -403,9 +403,16 @@ class AccountController extends Controller
   }
 
   public function send_email_bulk(Request $request){
+    if(!isset($request->accountid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih akun terlebih dahulu';
+      return $arr;
+    }
+
     $validator = $this->validator($request->all());
 
     if(!$validator->fails()){
+
       Mail::to($request->email)->queue(new ProfileBulkEmail($request->email,$request->type,$request->accountid));
 
       $arr['status'] = 'success';
@@ -431,7 +438,20 @@ class AccountController extends Controller
 
   public function add_groups(Request $request){
     $arr['status'] = 'success';
-    $arr['message'] = '';
+    $arr['message'] = 'Add to group berhasil';
+
+    //Pengecekkan id 
+    if(!isset($request->accountid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih akun terlebih dahulu';
+
+      return $arr;
+    } else if(!isset($request->groupid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih group terlebih dahulu';
+
+      return $arr;
+    }
 
     foreach ($request->accountid as $accountid) {
       foreach ($request->groupid as $groupid) {
@@ -467,6 +487,14 @@ class AccountController extends Controller
   }
 
   public function save_groups(Request $request){
+    //Pengecekkan id 
+    if(!isset($request->accountid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih akun terlebih dahulu';
+
+      return $arr;
+    }
+
     foreach ($request->accountid as $accountid) {
       $checksave = Save::where('user_id',Auth::user()->id)
                       ->where('account_id',$accountid)
@@ -493,6 +521,14 @@ class AccountController extends Controller
     $arr['status'] = 'success';
     $arr['message'] = '';
 
+    //Pengecekkan id 
+    if(!isset($request->historyid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih history terlebih dahulu';
+
+      return $arr;
+    }
+
     foreach ($request->historyid as $id) {
       $history = HistorySearch::find($id)->delete(); 
     }
@@ -503,6 +539,12 @@ class AccountController extends Controller
   public function print_pdf_bulk(Request $request){
     if(Auth::user()->membership=='free' or Auth::user()->membership=='pro'){
       return abort(403);
+    }
+
+    if(!isset($request->accountid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih akun terlebih dahulu';
+      return $arr;
     }
 
     $user = User::find(Auth::user()->id);
@@ -543,6 +585,12 @@ class AccountController extends Controller
 
     if(Auth::user()->membership=='free' or Auth::user()->membership=='pro'){
       return abort(403);
+    }
+
+    if(!isset($request->accountid)){
+      $arr['status'] = 'error';
+      $arr['message'] = 'Pilih akun terlebih dahulu';
+      return $arr;
     }
 
     $user = User::find(Auth::user()->id);
